+++ src/pages/Home.js
@@ -1,7 +1,4 @@
-import { postLogin } from '../utils/axios-fn';
-
 function Home() {
-  postLogin();
   return <div>Home</div>;
 }
